Add logic to resolve stream status
Logic has been added to resolve stream status & filter the stream
switcher accordingly. If a stream is fully offline, it is hidden from
the switcher; if a stream is "asleep" (playing automation), the name is
greyed out.
Other cleanup has been performed as part of this commit. Notably, we
gained the IScheduleState type as a reference for how part of the Redux
state should look, and the unused loadSchedule function was removed.