Fix potential site crash if the first stream has no slate
Previously it was assumed the first stream always carried the slate
object, but this turns out to not be true: the return order depends on
the last modified time. So instead I've written some logic to resolve
the first slate based on the highest online priority, but checking lower
priority streams in turn to have the greatest chance of always resolving
a slate. A side effect of this is that the selected stream always starts
off as the first one with a slate, regardless of its online status.