Add support for multiple streams (#73)
* Add player state + loading icon
* Basic containers using stream array
* Initial player experiments
* Improve Homepage Block styling
* Basic stream switcher support & fixed some player behaviours
* Fixed OneImage not updating with state changes & vertically centered play button
* Fixing schedule page
* Fixed homepage console warnings
* Fixed webpack config so that webpack doesn't load itself twice.
From the webpack documentation: "As a rule of thumb: Use exactly one entry point for each HTML document." Therefore I have removed the old `vendor` entrypoint and let webpack resolve dependencies for the two application entrypoints.

* Fix leaky component by using componentWillUnmount
LiveMinutesHigherOrder was leaking memory due to an uncleared timeout.
Fix leak by renaming erroneous componentDidLeave method to
componentWillUnmount.

* Enable hot reloading for the mini-css-extract plugin
* Replace select-based stream switcher with custom component
Rather than using the standard (but ugly and difficult to style) HTML
select element, instead use React components

* Fix an uncleared interval in the DateInterval component
* Improve styling of stream switcher
Added:
* Open/close animation
* Default entry description
* Arrow showing menu open state

* Reformat schedule query
* Add API_HOST constant
* Add logic to resolve stream status
Logic has been added to resolve stream status & filter the stream
switcher accordingly. If a stream is fully offline, it is hidden from
the switcher; if a stream is "asleep" (playing automation), the name is
greyed out.
Other cleanup has been performed as part of this commit. Notably, we
gained the IScheduleState type as a reference for how part of the Redux
state should look, and the unused loadSchedule function was removed.

* Tidy up switcher side-effects; consolidate effects into Loadable variant
- Side-effects (resolving the stream order) have been consolidated into
the LoadableStreamSwitcher component
- Loading text block is fixed; it now flexes correctly

* Add missing caret-up file
* Fix potential site crash if the first stream has no slate
Previously it was assumed the first stream always carried the slate
object, but this turns out to not be true: the return order depends on
the last modified time. So instead I've written some logic to resolve
the first slate based on the highest online priority, but checking lower
priority streams in turn to have the greatest chance of always resolving
a slate. A side effect of this is that the selected stream always starts
off as the first one with a slate, regardless of its online status.