Fix initial stream sort to resolve in the same order as the later sort
Previously the initial sort was sorting highest-first, while the later
(full) sort is lowest-first. This came about because I wrote the full
sort code while tired, and accidentally wrote code for lowest-first.
This temporarily broke the website (in the eyes of users) as the stream
intended to be the lowest priority ended up being selected by default.
When I hotfixed that by simply inverting all the priorities, I decided
that I liked the lowest-first ordering better (as it puts our main
stream at priority 1). However, as a result of all this, this other
sort call was the wrong way round. So that's fixed here.

The real fix for this situation is refactoring the rest of the code
to handle a null slate.