Fix schedule timezone wackiness
Now that Europe/London is an hour ahead of UTC, a bit of code that
previously appeared to be working revealed itself to be incorrect -
it was interpreting a UTC timestamp as if it was Europe/London, which
was effectively a pass-through until the two timezones stopped being
the same. The at() function previously used is designed for parsing
times-of-day received from the API, since those are always in
Europe/London.