Fix DST-caused wackiness in schedule
Finally tracked this down! The dayjs#diff function was correctly
figuring out the difference between two times on a sunday to be 12
hours, despite there only being 11 hours in a naive comparison (00:00 to
11:00am); because it's DST changeover weekend, 02:00 is repeated, so
there's an extra hour! Fix by adding a diffing function that takes this
into account and applies an offset to the duration.
NB: lord knows what would happen if we actually tried to have a show
between 02:00 and 03:00 on a Sunday.